requirements.txt
django==4.2.4
djangorestframework==3.14.0
pdfkit==1.0.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-rest-passwordreset==1.3.0
facebook-sdk==3.1.0
google-auth==2.22.0
